réorganisation controller et ajout sécurité all routes

+++ fngcine/src/main/java/com/boutiqueEnLigne/fngcine/controller/AdminController.java
@@ -1,20 +1,19 @@
 package com.boutiqueEnLigne.fngcine.controller;
 
-import com.boutiqueEnLigne.fngcine.entity.Order;
 import com.boutiqueEnLigne.fngcine.entity.Product;
-import com.boutiqueEnLigne.fngcine.entity.User;
+import com.boutiqueEnLigne.fngcine.service.OrderDetailService;
 import com.boutiqueEnLigne.fngcine.service.OrderService;
 import com.boutiqueEnLigne.fngcine.service.ProductService;
 import com.boutiqueEnLigne.fngcine.service.UserService;
+import com.boutiqueEnLigne.fngcine.validation.AuthentificationValidation;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @Getter
@@ -30,61 +29,41 @@ public class AdminController {
     @Autowired
     private ProductService productService;
 
-    // -------------------------------- USER -------------------------------------- //
+    @Autowired
+    private OrderDetailService orderDetailService;
 
-    @GetMapping("/users")
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<Iterable<User>> getUsers() {
-        Iterable<User> userList = userService.getUsers();
-        return ResponseEntity.ok(userList);
-    }
-
-    @PutMapping("/users/{userId}")
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<User> updateUser(@PathVariable("userId") Long userId, @RequestBody User userToUpdate){
-        Long id = userToUpdate.getId();
-        User user = userService.getUserById(id);
-        String password = user.getPassword();
-        userToUpdate.setPassword(password);
-        User updatedUser = userService.updateUser(userToUpdate);
-        return new ResponseEntity<>(updatedUser, HttpStatus.OK);
-    }
-
-    @DeleteMapping("/users/{userId}")
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public void deleteUserById(@PathVariable("userId") Long id) {
-        User user = userService.getUserById(id);
-        userService.deleteUserById(id);
-    }
-
-    // -------------------------------- ORDER -------------------------------------- //
-
-    @GetMapping("/orders")
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<List<Order>> getOrders() {
-        List<Order> orderList = orderService.getOrders();
-        return ResponseEntity.ok(orderList);
-    }
-
-    // -------------------------------- PRODUCT -------------------------------------- //
+    // -------------------------------- ADMIN MANAGE PRODUCTS -------------------------------------- //
 
     @PostMapping("/products")
-    public ResponseEntity<Product> createProduct(@RequestBody Product product) {
-        Product newProduct = productService.createProduct(product);
-        return new ResponseEntity<>(newProduct, HttpStatus.CREATED);
+    @PreAuthorize("hasAuthority('ADMIN')")
+    public ResponseEntity<Product> createProduct(@RequestBody Product product, AuthentificationValidation authentificationValidation) {
+        if(authentificationValidation.isAdmin()){
+            Product newProduct = productService.createProduct(product);
+            return new ResponseEntity<>(newProduct, HttpStatus.CREATED);
+        }else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 
     @PutMapping("/products/{id}")
     @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<Product> updateProduct(@PathVariable("id") Long id, @RequestBody Product productToUpdate){
+    public ResponseEntity<Product> updateProduct(@PathVariable("id") Long id, @RequestBody Product productToUpdate, AuthentificationValidation authentificationValidation){
+        if(authentificationValidation.isAdmin()){
         Product updatedProduct = productService.updateProduct(productToUpdate);
         return new ResponseEntity<>(updatedProduct, HttpStatus.OK);
+        } else {
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
     }
 
     @DeleteMapping("/products/{id}")
     @PreAuthorize("hasAuthority('ADMIN')")
-    public void deleteProductById(@PathVariable("id") Long id) {
+    public void deleteProductById(@PathVariable("id") Long id, AuthentificationValidation authentificationValidation) {
+        if(authentificationValidation.isAdmin()){
         productService.deleteProduct(id);
+        } else {
+           ResponseEntity.status(HttpStatus.FORBIDDEN);
+        }
     }
 
 }
